Models That Are Less Restricted

Generate a list of models with one or more fixed parameter freed.

Usage

get_add(
  sem_out,
  must_add = NULL,
  must_not_add = NULL,
  remove_constraints = TRUE,
  exclude_error_cov = TRUE,
  df_change = 1,
  model_id = NA,
  keep_correct_df_change = TRUE
)

Arguments

sem_out

The output from an structural equation modeling function. Currently support lavaan::lavaan only. Usually the one used in get_add() or get_drop() to generate the list of models.

must_add

A character vector of parameters, named in lavaan::lavaan() style (e.g., "y ~ x"), that must be added. Default is `NULL``.

must_not_add

A character vector of parameters, named in lavaan::lavaan() style (e.g., "x1 ~~ x1"), that must not be added. Default is NULL.

remove_constraints

Whether equality constraints will be removed. Default is ``TRUE`.

exclude_error_cov

Exclude error covariances of indicators. Default is TRUE.

df_change

How many degrees of freedom (df) away in the list. All models with df change less than or equal to this number will be included. Default is 1.

model_id

The identification number of the starting model. Default is NA, no identification number.

keep_correct_df_change

Keep only tables with actual df change equal to expected df change.

Value

An object of the class partables, a named list of parameter tables, each of them to be used by lavaan::lavaan() or update()

for fitting a model with the added parameters.

Details

Generate a list of models with one or more fixed parameter freed (and the degrees of freedom, df, increases by one or more). If a model has equality constraints, models with one or more of the constraints released will also be included.

Conceptually, paths or covariances are "added" to form the list of models.

The models to be included are identified by lavaan::modificationIndices().

The models will be checked by lavaan to make sure that the increase in model degrees of freedom is of the expected value.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
dat <- dat_path_model
mod <-
"
x3 ~ a*x1 + b*x2
x4 ~ a*x1
ab := a*b
"
fit <- sem(mod, dat_path_model, fixed.x = TRUE)
mod_to_add <- get_add(fit)
mod_to_add
#> 
#> Call:
#> get_add(sem_out = fit)
#> 
#> Number of parameter tables: 2
#> 
#> The modifications:
#> add: x4~x2
#> add: (x3~x1),(x4~x1)